reorganized validation arrays into validations.js
