+++ routes/users.js
@@ -7,6 +7,7 @@ const db = require('../db/models');
 const { csrfProtection, asyncHandler } = require('./utils');
 
 const { loginUser, logoutUser, requireAuth } = require('../auth');
+const { userValidators, loginValidators } = require('../validations');
 
 const router = express.Router();
 
@@ -18,47 +19,6 @@ router.get('/sign-up', csrfProtection, (req, res) => {
     csrfToken: req.csrfToken(),
   });
 });
-
-const userValidators = [
-  check('username')
-    .exists({ checkFalsy: true })
-    .withMessage('Please provide a value for Username')
-    .isLength({ max: 25 })
-    .withMessage('Username must not be more than 25 characters long'),
-  check('email')
-    .exists({ checkFalsy: true })
-    .withMessage('Please provide a value for Email Address')
-    .isLength({ max: 255 })
-    .withMessage('Email Address must not be more than 255 characters long')
-    .isEmail()
-    .withMessage('Email Address is not a valid email')
-    .custom((value) => {
-      return db.User.findOne({ where: { email: value } })
-        .then((user) => {
-          if (user) {
-            return Promise.reject('The provided Email Address is already in use by another account');
-          }
-        });
-    }),
-  check('password')
-    .exists({ checkFalsy: true })
-    .withMessage('Please provide a value for Password')
-    .isLength({ max: 50 })
-    .withMessage('Password must not be more than 50 characters long')
-    .matches(/^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*])/, 'g')
-    .withMessage('Password must contain at least 1 lowercase letter, uppercase letter, number, and special character (i.e. "!@#$%^&*")'),
-    check('confirmPassword')
-    .exists({ checkFalsy: true })
-    .withMessage('Please provide a value for Confirm Password')
-    .isLength({ max: 50 })
-    .withMessage('Confirm Password must not be more than 50 characters long')
-    .custom((value, { req }) => {
-      if (value !== req.body.password) {
-        throw new Error('Confirm Password does not match Password');
-      }
-      return true;
-    }),
-];
 
 router.post('/sign-up', csrfProtection, userValidators,
   asyncHandler(async (req, res) => {
@@ -100,15 +60,6 @@ router.get('/log-in', csrfProtection, (req, res) => {
     csrfToken: req.csrfToken(),
   });
 });
-
-const loginValidators = [
-  check('email')
-    .exists({ checkFalsy: true })
-    .withMessage('Please provide a value for Email Address'),
-  check('password')
-    .exists({ checkFalsy: true })
-    .withMessage('Please provide a value for Password'),
-];
 
 router.post('/log-in', csrfProtection, loginValidators,
   asyncHandler(async (req, res) => {
